requirements.txt
discord.py
numpy
Pillow
aiofiles